removed duplication of core files
