+++ configs/templates/component.yaml
