+++ lib/wraith/javascript/beforeCapture.js
